update skeet init/ai/check
