--- packages/cli/src/cli/init/initCloud/initWhenGithubActionsCreated.ts
@@ -1,7 +1,0 @@
-import { setupNetwork } from '@/lib'
-import { updateSkeetCloudConfigCloudStatus } from '../updateSkeetCloudConfigCloudStatus'
-
-export const initWhenGithubActionsCreated = async () => {
-  await setupNetwork()
-  await updateSkeetCloudConfigCloudStatus('VPN_CREATED')
-}
