--- packages/cli/src/cli/init/initCloud/initWhenSQLCreated.ts
@@ -1,5 +1,0 @@
-import chalk from 'chalk'
-
-export const initWhenSQLCreated = async () => {
-  console.log(chalk.white('Coming soon!'))
-}
