+++ packages/cli/src/cli/init/initStep/createVPN.ts
@@ -1,0 +1,7 @@
+import { setupNetwork } from '@/lib'
+import { updateSkeetCloudConfigCloudStatus } from '../../init/updateSkeetCloudConfigCloudStatus'
+
+export const createVPN = async () => {
+  await setupNetwork()
+  await updateSkeetCloudConfigCloudStatus('VPN_CREATED')
+}
